+++ src/decomp/sas_c/newgrid_update_selection_from_input_alt.c
@@ -30,7 +30,7 @@ extern UBYTE TEXTDISP_PrimaryGroupPresentFlag;
 extern LONG CONFIG_TimeWindowMinutes;
 
 extern void NEWGRID_ClearMarkersIfSelectable(LONG mode, LONG row);
-extern LONG NEWGRID_UpdatePresetEntry(char **entryPtr, char **auxPtr, LONG row, LONG col);
+extern short NEWGRID_UpdatePresetEntry(char **entryPtr, char **auxPtr, LONG row, LONG col);
 extern LONG NEWGRID_TestEntrySelectable(const void *entry, const void *aux, LONG mode);
 extern LONG NEWGRID2_JMPTBL_DISPLIB_FindPreviousValidEntryIndex(const char *entry, const char *aux, LONG idx);
 extern LONG NEWGRID2_JMPTBL_ESQ_TestBit1Based(const UBYTE *bitset, LONG idx);
